Embedding echosign docs with iframes
Addresses:
- [WEB-472]
- [WEB-474]
- [WEB-476]

Added better instructional text on both pages with embeddings,
explaining what to do if there is no Flash, and how to get to the direct
signing page if the embedding doesn't load.

Switched from javascript embed (not loading on localhost or staging) to
an iframe embed, which loaded in all environments.